backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
requests==2.32.3
wafw00f==2.2.0
beautifulsoup4==4.12.3
